Reorganized components into folders based on their route, event listners in gameroom component did mount
